Begin writing test cases for validating nrrdread
Fixed issue with reading strings with spaces in the words, NRRD uses quotations in some of the array of strings options. This includes space units, units and labels that have quotations around each word.

Change FlipDomain to FlipAxes that switches the ordering of all of the dimensions. This is because MATLAB uses Fortran ordering which completely reverses all of the axes.