Fix floating point precision in nrrdwrite
Previously, only 5 floating points would be shown when writing to a file which is way below the precision of a double. Also, regardless of the number of significant digits, the number of digits written was always 5.

Now, only the necessary number of digits are written and a maximum of 16 digits is written because this is the smallest precision of a 64-bit double.