Remove FlipAxes parameter
I had a fundamental misunderstanding of the difference between Fortran and C-order style. I was using permute to reverse all of the dimension locations. Rather, it should only be performed on the horizontal and vertical dimensions. Since this can be subjective based on the data being used, the parameter was removed and left as a task for the user to perform.

Added a note in REMARKS about using permute to fix the orientation of any images being loaded from NRRD.